Split off modification cache stuff into separate file, get things to work on 32 bit.
